issue #336: file-backed mmap の file offset も long 化 (alloc_and_map / AllocInfo.map_offset)
#336 の残スコープ。mmap の off_t は FileSeek 手前の alloc_and_map(int offset) /
AllocInfo.map_offset(int) で既に切り詰められていた (file-backed mmap を >2GB offset で
張ると file[0..] を誤って map してしまう)。

- MemoryBackend.alloc_and_map の 3 overload の offset を long に。
- Memory / NativeMemoryBackend の alloc_and_map 実装と AllocInfo.map_offset を long に。
- amd64_mmap の (int)offset cast を除去。
- size / map_size は byte[] の 2GB 上限ゆえ据置、srcoff は region 内 offset (<2GB) ゆえ
  据置、i386 sys_mmap は int offset を auto-promote (32-bit ABI、無改変)。

回帰: 新規 sys_mmap_large64 (4GiB offset の file-backed mmap、byte0='M' を確認、sparse)
が software / native(KVM) 両方で PASS。hello_dyn64 (ld.so の libc mmap) native PASS。
これで #336 完了 (lseek/pread64/pwrite64/copy_file_range + file-backed mmap offset)。

+++ src/main/java/emulin/NativeMemoryBackend.java
@@ -512,11 +512,11 @@ public final class NativeMemoryBackend implements MemoryBackend {
   // 旧シグネチャ経路 (mremap の addr=0 / i386 等、flags 情報が無い caller) は従来挙動 (addr!=0 を
   //   MAP_FIXED 扱い) を維持する。flags を知る amd64_mmap は 6 引数 alloc_and_map 経由で hint を渡す。
   @Override public long    alloc( long adrs, int size ) { return anonMmap( adrs, size, true ); }
-  @Override public long    alloc_and_map( long adrs, int size, int fd, int offset ) { return alloc_and_map( adrs, size, fd, offset, 0 ); }
-  @Override public long    alloc_and_map( long adrs, int size, int fd, int offset, int prot ) {
+  @Override public long    alloc_and_map( long adrs, int size, int fd, long offset ) { return alloc_and_map( adrs, size, fd, offset, 0 ); }
+  @Override public long    alloc_and_map( long adrs, int size, int fd, long offset, int prot ) {
     return alloc_and_map( adrs, size, fd, offset, prot, 0x10 /* MAP_FIXED 相当 = 従来挙動 */ );
   }
-  @Override public long    alloc_and_map( long adrs, int size, int fd, int offset, int prot, long flags ) {
+  @Override public long    alloc_and_map( long adrs, int size, int fd, long offset, int prot, long flags ) {
     long va = anonMmap( adrs, size, ( flags & 0x10 ) != 0 );   // 0x10 = MAP_FIXED (無し = adrs は hint)
     if( fd >= 0 ) {
       // file-backed mmap (ld.so が libc.so 等を map): file の [offset, offset+size) を guest に読む。
